+++ src/main/java/dto/LoginDTO.java
@@ -1,12 +1,9 @@
-package gift.controller;
+package dto;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 
-public record UserDTO(
-    @NotBlank(message = "이름을 입력하세요.")
-    String name,
-
+public record LoginDTO(
     @NotBlank(message = "이메일을 입력하세요.")
     @Email(message = "유효한 이메일을 입력하세요.")
     String email,
