+++ src/main/java/dto/ProductDTO.java
@@ -1,4 +1,4 @@
-package gift.controller;
+package dto;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Pattern;
