+++ src/main/java/dto/UserDTO.java
@@ -1,9 +1,12 @@
-package gift.controller;
+package dto;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 
-public record LoginDTO(
+public record UserDTO(
+    @NotBlank(message = "이름을 입력하세요.")
+    String name,
+
     @NotBlank(message = "이메일을 입력하세요.")
     @Email(message = "유효한 이메일을 입력하세요.")
     String email,
