+++ src/main/java/dto/WishlistRequest.java
@@ -1,4 +1,4 @@
-package gift.controller;
+package dto;
 
 import jakarta.validation.constraints.NotNull;
